Separate public machine contracts
