+++ src/internal/machine/symbols.ts
@@ -1,0 +1,2 @@
+/** @internal */
+export const InitialEventTypeId: unique symbol = Symbol("effect/Machine/InitialEvent")
