+++ src/unstable/reactivity/AtomMachine.ts
@@ -4,17 +4,17 @@
  * @since 4.0.0
  */
 
-import * as Data from "effect/Data"
-import * as Effect from "effect/Effect"
-import * as Equal from "effect/Equal"
-import * as Option from "effect/Option"
+import type * as Effect from "effect/Effect"
+import type * as Option from "effect/Option"
 import type * as Schema from "effect/Schema"
 import type * as Scope from "effect/Scope"
-import * as Stream from "effect/Stream"
-import { AsyncResult, Atom, type AtomRegistry } from "effect/unstable/reactivity"
-import * as Model from "../../internal/machine/model.js"
+import type * as Stream from "effect/Stream"
+import type { AsyncResult, Atom, AtomRegistry } from "effect/unstable/reactivity"
+import * as internal from "../../internal/machine/atom.js"
+import type { ChildNotActiveError, NotReadyError } from "../../internal/machine/atom.js"
+import type * as Model from "../../internal/machine/model.js"
 import type { EnsureExecutable } from "../../internal/machine/readiness.js"
-import * as Machine from "../../Machine.js"
+import type * as Machine from "../../Machine.js"
 
 /**
  * Error returned when a machine command is issued before startup completes.
@@ -22,7 +22,7 @@ import * as Machine from "../../Machine.js"
  * @category errors
  * @since 4.0.0
  */
-export class NotReadyError extends Data.TaggedError("NotReadyError") {}
+export { NotReadyError } from "../../internal/machine/atom.js"
 
 /**
  * Error returned when a command targets a child machine that is not active.
@@ -30,9 +30,7 @@ export class NotReadyError extends Data.TaggedError("NotReadyError") {}
  * @category errors
  * @since 4.0.0
  */
-export class ChildNotActiveError extends Data.TaggedError("ChildNotActiveError")<{
-  readonly id: string
-}> {}
+export { ChildNotActiveError } from "../../internal/machine/atom.js"
 
 type AtomSupportedRequirements = Scope.Scope | AtomRegistry.AtomRegistry
 
@@ -81,68 +79,6 @@ type MachineStartError<InitialE, E, InitialR, R, RuntimeError = never> =
   | Machine.StartupError
   | Machine.StoppedError
   | RuntimeError
-
-const runMachineAtomEffect = <State, Event, Error, Output, StartError, Requirements>(
-  get: Atom.AtomContext,
-  start: Effect.Effect<Machine.MachineRef<State, Event, Error, Output>, StartError, Requirements>
-): Effect.Effect<never, StartError, Requirements> =>
-  Effect.scoped(
-    Effect.acquireRelease(start, (ref) => ref.stop).pipe(
-      Effect.tap((ref) => Effect.sync(() => get.setSelf(AsyncResult.success(ref)))),
-      Effect.flatMap(() => Effect.never)
-    )
-  )
-
-const startMachineAtomEffect = <
-  const States extends Machine.Machine.StateSchemas,
-  const Events extends ReadonlyArray<Machine.Machine.TaggedSchema>,
-  const Emits extends ReadonlyArray<Machine.Machine.TaggedSchema> = any,
-  const Input extends Schema.Top = typeof Schema.Void,
-  UnhandledStates extends Machine.Machine.StateIdentifier<States> = Machine.Machine.StateIdentifier<States>,
-  E = never,
-  R = never,
-  InitialE = never,
-  InitialR = never,
-  FinalStates extends Machine.Machine.StateIdentifier<States> = never,
-  Output = never,
-  OutputStates extends Machine.Machine.StateIdentifier<States> = never,
-  InputEvents extends ReadonlyArray<Machine.Machine.TaggedSchema> = Events
->(
-  get: Atom.AtomContext,
-  machine:
-    & Machine.Machine<
-      States,
-      Events,
-      Input,
-      UnhandledStates,
-      E,
-      R,
-      InitialE,
-      InitialR,
-      FinalStates,
-      Output,
-      Emits,
-      OutputStates,
-      InputEvents
-    >
-    & EnsureExecutable<States, UnhandledStates, OutputStates>,
-  args: [...Machine.Machine.InputArgs<Input>]
-): Effect.Effect<
-  Machine.MachineRef<
-    Machine.Machine.Snapshot<States>,
-    Machine.Machine.EventOf<InputEvents>,
-    MachineRuntimeError<E, R>,
-    Output
-  >,
-  MachineStartError<InitialE, E, InitialR, R>,
-  MachineRequirements<InitialR, R, Machine.Machine.EventOf<Events>, Machine.Machine.EmitOf<Emits>>
-> => runMachineAtomEffect(get, Machine.start(machine, ...args))
-
-const resumeMachineAtomEffect = (
-  get: Atom.AtomContext,
-  machine: Machine.Machine.Any,
-  snapshot: Machine.Machine.Snapshot<any>
-) => runMachineAtomEffect(get, Machine.resume(machine as any, snapshot as any))
 
 /**
  * Atoms backed by one running machine instance in an `AtomRegistry`.
@@ -226,7 +162,9 @@ export interface MachineAtom<State, Event, Error = never, Output = never, StartE
 }
 
 type RefState<Ref> = Ref extends Machine.MachineRef<infer State, any, any, any> ? State : never
+
 type RefError<Ref> = Ref extends Machine.MachineRef<any, any, infer Error, any> ? Error : never
+
 type RefOutput<Ref> = Ref extends Machine.MachineRef<any, any, any, infer Output> ? Output : never
 
 /**
@@ -345,288 +283,6 @@ export type ChildOf<
   Child extends Machine.ChildMachine.Any
 > = ChildMachineAtom<Child, BridgeStartError<Parent>>
 
-const makeRuntimeResultAtom = <State, Error, Output, StartError>(
-  snapshot: Atom.Atom<AsyncResult.AsyncResult<Machine.RuntimeSnapshot<State, Error, Output>, StartError>>
-): Atom.Atom<AsyncResult.AsyncResult<State, StartError | Error>> =>
-  Atom.readable((get): AsyncResult.AsyncResult<State, StartError | Error> => {
-    const current = get(snapshot)
-    if (AsyncResult.isInitial(current)) {
-      return AsyncResult.initial(current.waiting)
-    } else if (AsyncResult.isFailure(current)) {
-      return AsyncResult.failureWithPrevious(current.cause, {
-        previous: get.self(),
-        waiting: current.waiting
-      })
-    } else if (current.value.status === "error") {
-      return AsyncResult.failureWithPrevious(current.value.cause, {
-        previous: get.self(),
-        waiting: current.waiting
-      })
-    }
-    return AsyncResult.success(current.value.state, {
-      waiting: current.waiting
-    })
-  }).pipe(Atom.withEquality(Equal.equals))
-
-const makeChildRuntimeResultAtom = <State, Error, Output, StartError>(
-  snapshot: Atom.Atom<
-    AsyncResult.AsyncResult<Option.Option<Machine.RuntimeSnapshot<State, Error, Output>>, StartError>
-  >
-): Atom.Atom<AsyncResult.AsyncResult<Option.Option<State>, StartError | Error>> =>
-  Atom.readable((get): AsyncResult.AsyncResult<Option.Option<State>, StartError | Error> => {
-    const current = get(snapshot)
-    if (AsyncResult.isInitial(current)) {
-      return AsyncResult.initial(current.waiting)
-    } else if (AsyncResult.isFailure(current)) {
-      return AsyncResult.failureWithPrevious(current.cause, {
-        previous: get.self(),
-        waiting: current.waiting
-      })
-    } else if (Option.isNone(current.value)) {
-      const previous = get.self<AsyncResult.AsyncResult<Option.Option<State>, StartError | Error>>()
-      if (Option.isSome(previous) && AsyncResult.isFailure(previous.value)) {
-        return previous.value
-      }
-      return AsyncResult.success(Option.none(), {
-        waiting: current.waiting
-      })
-    } else if (current.value.value.status === "error") {
-      return AsyncResult.failureWithPrevious(current.value.value.cause, {
-        previous: get.self(),
-        waiting: current.waiting
-      })
-    }
-    return AsyncResult.success(Option.some(current.value.value.state), {
-      waiting: current.waiting
-    })
-  }).pipe(Atom.withEquality(Equal.equals))
-
-const makeChildRefAtom = <Child extends Machine.ChildMachine.Any, StartError>(
-  parentRef: Atom.Atom<
-    AsyncResult.AsyncResult<Option.Option<Machine.MachineRef<any, any, any, any>>, StartError>
-  >,
-  child: Child
-): Atom.Atom<AsyncResult.AsyncResult<Option.Option<Machine.ChildMachine.Ref<Child>>, StartError>> =>
-  Atom.readable((get) => {
-    const parent = get(parentRef)
-    if (AsyncResult.isInitial(parent)) {
-      return AsyncResult.initial(parent.waiting)
-    } else if (AsyncResult.isFailure(parent)) {
-      return AsyncResult.failureWithPrevious(parent.cause, {
-        previous: get.self(),
-        waiting: parent.waiting
-      })
-    } else if (Option.isNone(parent.value)) {
-      return AsyncResult.success(Option.none())
-    }
-
-    const handle = parent.value.value
-    const current = Effect.runSync(handle.child(child))
-    const cancel = Effect.runCallback(
-      Effect.yieldNow.pipe(
-        Effect.andThen(
-          handle.childChanges(child).pipe(
-            Stream.runForEach((ref) => Effect.sync(() => get.setSelf(AsyncResult.success(ref))))
-          )
-        )
-      )
-    )
-    get.addFinalizer(cancel)
-    return AsyncResult.success(current)
-  })
-
-const makeChildFromRefAtom = <Child extends Machine.ChildMachine.Any, StartError>(
-  ref: Atom.Atom<AsyncResult.AsyncResult<Option.Option<Machine.ChildMachine.Ref<Child>>, StartError>>,
-  descriptor: Child
-): ChildMachineAtom<Child, StartError> => {
-  type Ref = Machine.ChildMachine.Ref<Child>
-  type State = RefState<Ref>
-  type Error = RefError<Ref>
-  type Output = RefOutput<Ref>
-
-  const snapshot = Atom.readable((get): AsyncResult.AsyncResult<
-    Option.Option<Machine.RuntimeSnapshot<State, Error, Output>>,
-    StartError
-  > => {
-    const result = get(ref)
-    if (AsyncResult.isInitial(result)) {
-      return AsyncResult.initial(result.waiting)
-    } else if (AsyncResult.isFailure(result)) {
-      return AsyncResult.failureWithPrevious(result.cause, {
-        previous: get.self(),
-        waiting: result.waiting
-      })
-    } else if (Option.isNone(result.value)) {
-      return AsyncResult.success(Option.none())
-    }
-
-    const handle = result.value.value as unknown as Machine.MachineRef<State, any, Error, Output>
-    const cancel = Effect.runCallback(
-      handle.changes.pipe(
-        Stream.runForEach((snapshot) => Effect.sync(() => get.setSelf(AsyncResult.success(Option.some(snapshot)))))
-      )
-    )
-    get.addFinalizer(cancel)
-    return AsyncResult.success(Option.some(Effect.runSync(handle.snapshot)))
-  })
-
-  const send = Atom.writable<
-    AsyncResult.AsyncResult<void, StartError | NotReadyError | ChildNotActiveError | Machine.StoppedError>,
-    Machine.ChildMachine.Event<Child>
-  >(
-    (get) => AsyncResult.map(get(ref), () => undefined),
-    (ctx, event) => {
-      const result = ctx.get(ref)
-      if (AsyncResult.isInitial(result)) {
-        ctx.setSelf(AsyncResult.fail(new NotReadyError()))
-      } else if (AsyncResult.isFailure(result)) {
-        ctx.setSelf(AsyncResult.map(result, () => undefined))
-      } else if (Option.isNone(result.value)) {
-        ctx.setSelf(AsyncResult.fail(new ChildNotActiveError({ id: descriptor.id })))
-      } else {
-        Effect.runCallback(result.value.value.send(event as never), {
-          onExit: (exit) => ctx.setSelf(AsyncResult.fromExit(exit))
-        })
-      }
-    }
-  )
-
-  const stop = Atom.writable<
-    AsyncResult.AsyncResult<void, StartError | NotReadyError | ChildNotActiveError>,
-    void
-  >(
-    (get) => AsyncResult.map(get(ref), () => undefined),
-    (ctx) => {
-      const result = ctx.get(ref)
-      if (AsyncResult.isInitial(result)) {
-        ctx.setSelf(AsyncResult.fail(new NotReadyError()))
-      } else if (AsyncResult.isFailure(result)) {
-        ctx.setSelf(AsyncResult.map(result, () => undefined))
-      } else if (Option.isNone(result.value)) {
-        ctx.setSelf(AsyncResult.fail(new ChildNotActiveError({ id: descriptor.id })))
-      } else {
-        Effect.runCallback(result.value.value.stop)
-      }
-    }
-  )
-
-  const childFamily = Atom.family((nested: Machine.ChildMachine.Any) =>
-    makeChildFromRefAtom(
-      makeChildRefAtom(ref as any, nested),
-      nested
-    )
-  )
-  const child = <Nested extends Machine.ChildMachine.Any>(
-    nested: Nested
-  ): ChildMachineAtom<Nested, StartError> => childFamily(nested) as ChildMachineAtom<Nested, StartError>
-
-  return {
-    ref,
-    snapshot,
-    state: Atom.mapResult(snapshot, Option.map((snapshot) => snapshot.state)),
-    result: makeChildRuntimeResultAtom(snapshot),
-    send,
-    stop,
-    child
-  }
-}
-
-const makeFromRefAtom = <State, Event, Error, Output, StartError>(
-  ref: Atom.Atom<AsyncResult.AsyncResult<Machine.MachineRef<State, Event, Error, Output>, StartError>>
-): MachineAtom<State, Event, Error, Output, StartError> => {
-  const snapshot = Atom.readable((
-    get
-  ): AsyncResult.AsyncResult<Machine.RuntimeSnapshot<State, Error, Output>, StartError> => {
-    const result = get(ref)
-    if (AsyncResult.isInitial(result)) {
-      return AsyncResult.initial(result.waiting)
-    } else if (AsyncResult.isFailure(result)) {
-      return AsyncResult.failureWithPrevious(result.cause, {
-        previous: get.self<AsyncResult.AsyncResult<Machine.RuntimeSnapshot<State, Error, Output>, StartError>>(),
-        waiting: result.waiting
-      })
-    }
-
-    const handle = result.value
-    const cancel = Effect.runCallback(
-      handle.changes.pipe(
-        Stream.runForEach((snapshot) =>
-          Effect.sync(() =>
-            get.setSelf(
-              AsyncResult.success(snapshot, {
-                waiting: snapshot.status === "active"
-              })
-            )
-          )
-        )
-      )
-    )
-    get.addFinalizer(cancel)
-
-    const current = Effect.runSync(handle.snapshot)
-    return AsyncResult.success(current, {
-      waiting: current.status === "active"
-    })
-  })
-
-  const send = Atom.writable<
-    AsyncResult.AsyncResult<void, StartError | NotReadyError | Machine.StoppedError>,
-    Event
-  >(
-    (get) => AsyncResult.map(get(ref), () => undefined),
-    (ctx, event: Event) => {
-      const result = ctx.get(ref)
-      if (AsyncResult.isInitial(result)) {
-        ctx.setSelf(AsyncResult.fail(new NotReadyError()))
-      } else if (AsyncResult.isFailure(result)) {
-        ctx.setSelf(AsyncResult.map(result, () => undefined))
-      } else {
-        Effect.runCallback(result.value.send(event), {
-          onExit: (exit) =>
-            ctx.setSelf(
-              AsyncResult.fromExit(exit)
-            )
-        })
-      }
-    }
-  )
-
-  const stop = Atom.writable<AsyncResult.AsyncResult<void, StartError | NotReadyError>, void>(
-    (get) => AsyncResult.map(get(ref), () => undefined),
-    (ctx) => {
-      const result = ctx.get(ref)
-      if (AsyncResult.isInitial(result)) {
-        ctx.setSelf(AsyncResult.fail(new NotReadyError()))
-      } else if (AsyncResult.isFailure(result)) {
-        ctx.setSelf(AsyncResult.map(result, () => undefined))
-      } else {
-        Effect.runCallback(result.value.stop)
-      }
-    }
-  )
-
-  const optionalRef = Atom.mapResult(ref, Option.some)
-  const childFamily = Atom.family((descriptor: Machine.ChildMachine.Any) =>
-    makeChildFromRefAtom(
-      makeChildRefAtom(optionalRef as any, descriptor),
-      descriptor
-    )
-  )
-  const child = <Child extends Machine.ChildMachine.Any>(
-    descriptor: Child
-  ): ChildMachineAtom<Child, StartError> => childFamily(descriptor) as ChildMachineAtom<Child, StartError>
-
-  return {
-    ref,
-    snapshot,
-    state: Atom.mapResult(snapshot, (snapshot) => snapshot.state),
-    result: makeRuntimeResultAtom(snapshot),
-    send,
-    stop,
-    child
-  }
-}
-
 type SnapshotNode<State> = State extends Machine.Machine.AtomicSnapshot<string, unknown> ?
     | State
     | (State extends { readonly state: infer Child } ? SnapshotNode<Child>
@@ -643,17 +299,6 @@ type SnapshotValueByIdentifier<State, Path extends SnapshotIdentifier<State>> = 
   : never
 
 type ChildState<Child extends Machine.ChildMachine.Any> = RefState<Machine.ChildMachine.Ref<Child>>
-
-const selectSnapshot = <
-  State extends Machine.Machine.AtomicSnapshot<string, unknown>,
-  Path extends SnapshotIdentifier<State>
->(
-  snapshot: State,
-  path: Path
-): Option.Option<SnapshotValueByIdentifier<State, Path>> =>
-  Model.getSnapshotByPath(snapshot, path).pipe(
-    Option.map((snapshot) => snapshot.value)
-  ) as Option.Option<SnapshotValueByIdentifier<State, Path>>
 
 /**
  * Selects the typed value for an active state path.
@@ -672,22 +317,16 @@ const selectSnapshot = <
  * @category combinators
  * @since 4.0.0
  */
-export const select = <
+export const select: <
   State extends Machine.Machine.AtomicSnapshot<string, unknown>,
   Event,
   Error,
   Output,
   StartError,
   const Path extends SnapshotIdentifier<State>
->(
-  self: MachineAtom<State, Event, Error, Output, StartError>,
-  path: Path
-): Atom.Atom<
+>(self: MachineAtom<State, Event, Error, Output, StartError>, path: Path) => Atom.Atom<
   AsyncResult.AsyncResult<Option.Option<SnapshotValueByIdentifier<State, Path>>, StartError | Error>
-> =>
-  Atom.mapResult(self.result, (snapshot) => selectSnapshot(snapshot, path)).pipe(
-    Atom.withEquality(Equal.equals)
-  )
+> = internal.select
 
 /**
  * Selects the typed value for an active state path in an invoked child.
@@ -706,23 +345,16 @@ export const select = <
  * @category combinators
  * @since 4.0.0
  */
-export const selectChild = <
+export const selectChild: <
   Child extends Machine.ChildMachine.Any,
   StartError,
   const Path extends SnapshotIdentifier<ChildState<Child>>
->(
-  self: ChildMachineAtom<Child, StartError>,
-  path: Path
-): Atom.Atom<
+>(self: ChildMachineAtom<Child, StartError>, path: Path) => Atom.Atom<
   AsyncResult.AsyncResult<
     Option.Option<SnapshotValueByIdentifier<ChildState<Child>, Path>>,
     StartError | RefError<Machine.ChildMachine.Ref<Child>>
   >
-> =>
-  Atom.mapResult(
-    self.result,
-    Option.flatMap((snapshot) => selectSnapshot(snapshot, path))
-  ).pipe(Atom.withEquality(Equal.equals))
+> = internal.selectChild
 
 /**
  * Returns whether a state path is active.
@@ -734,7 +366,7 @@ export const selectChild = <
  * @category combinators
  * @since 4.0.0
  */
-export const matches = <
+export const matches: <
   State extends Machine.Machine.AtomicSnapshot<string, unknown>,
   Event,
   Error,
@@ -744,10 +376,7 @@ export const matches = <
 >(
   self: MachineAtom<State, Event, Error, Output, StartError>,
   path: Path
-): Atom.Atom<AsyncResult.AsyncResult<boolean, StartError | Error>> =>
-  Atom.mapResult(self.result, (snapshot) => Option.isSome(Model.getSnapshotByPath(snapshot, path))).pipe(
-    Atom.withEquality(Equal.equals)
-  )
+) => Atom.Atom<AsyncResult.AsyncResult<boolean, StartError | Error>> = internal.matches
 
 /**
  * Returns whether a state path is active in an invoked child.
@@ -759,20 +388,13 @@ export const matches = <
  * @category combinators
  * @since 4.0.0
  */
-export const matchesChild = <
+export const matchesChild: <
   Child extends Machine.ChildMachine.Any,
   StartError,
   const Path extends SnapshotIdentifier<ChildState<Child>>
->(
-  self: ChildMachineAtom<Child, StartError>,
-  path: Path
-): Atom.Atom<
+>(self: ChildMachineAtom<Child, StartError>, path: Path) => Atom.Atom<
   AsyncResult.AsyncResult<boolean, StartError | RefError<Machine.ChildMachine.Ref<Child>>>
-> =>
-  Atom.mapResult(
-    self.result,
-    Option.exists((snapshot) => Option.isSome(Model.getSnapshotByPath(snapshot, path)))
-  ).pipe(Atom.withEquality(Equal.equals))
+> = internal.matchesChild
 
 const BoundRequirementsTypeId = "~effect/reactivity/AtomMachine/BoundRequirements"
 
@@ -942,10 +564,7 @@ export const make: {
     Output,
     MachineStartError<InitialE, E, InitialR, R>
   >
-} = ((machine: Machine.Machine.Any, ...args: ReadonlyArray<unknown>) => {
-  const ref = Atom.make((get) => startMachineAtomEffect(get, machine as any, args as []))
-  return makeFromRefAtom(ref as any)
-}) as any
+} = internal.make
 
 /**
  * Creates a lazy atom bridge from a decoded logical snapshot.
@@ -965,28 +584,7 @@ export const resume: {
       & EnsureMachineExecutable<NoInfer<M>>,
     snapshot: Machine.Machine.Snapshot<Machine.Machine.States<M>>
   ): ResumedMachineAtomOf<M, never>
-} = ((machine: Machine.Machine.Any, snapshot: Machine.Machine.Snapshot<any>) => {
-  const ref = Atom.make((get) => resumeMachineAtomEffect(get, machine, snapshot))
-  return makeFromRefAtom(ref as any)
-}) as any
-
-const makeWithRuntime = (
-  runtime: Atom.AtomRuntime<any, any>,
-  machine: Machine.Machine.Any,
-  args: ReadonlyArray<unknown>
-): MachineAtom<any, any, any, any, any> => {
-  const ref = runtime.atom((get) => startMachineAtomEffect(get, machine as any, args as []))
-  return makeFromRefAtom(ref as any)
-}
-
-const resumeWithRuntime = (
-  runtime: Atom.AtomRuntime<any, any>,
-  machine: Machine.Machine.Any,
-  snapshot: Machine.Machine.Snapshot<any>
-): MachineAtom<any, any, any, any, any> => {
-  const ref = runtime.atom((get) => resumeMachineAtomEffect(get, machine, snapshot))
-  return makeFromRefAtom(ref as any)
-}
+} = internal.resume
 
 /**
  * Creates an `AtomMachine` factory that owns a shared Effect runtime.
@@ -998,16 +596,6 @@ const resumeWithRuntime = (
  * @category constructors
  * @since 4.0.0
  */
-export const bind = <Services, RuntimeError>(
+export const bind: <Services, RuntimeError>(
   runtime: Atom.AtomRuntime<Services, RuntimeError>
-): Bound<Services, RuntimeError> => ({
-  make:
-    ((machine: Machine.Machine.Any, ...args: ReadonlyArray<unknown>) =>
-      makeWithRuntime(runtime, machine, args)) as Bound<
-        Services,
-        RuntimeError
-      >["make"],
-  resume:
-    ((machine: Machine.Machine.Any, snapshot: Machine.Machine.Snapshot<any>) =>
-      resumeWithRuntime(runtime, machine, snapshot)) as Bound<Services, RuntimeError>["resume"]
-})
+) => Bound<Services, RuntimeError> = internal.bind
